semantic-kernel/code-samples/requirements.txt
semantic-kernel==1.28.0
python-dotenv==1.1.0